refactor net definitions
